Added files created in Summer 2018
These are most of the files I created during my summer research. The original files can be found on the UCSC graymalkin server. A lot of the files in the data folder may not be needed, I just copied the whole folder over to make sure that I did not forget anything.