Changing table to work by spaxel
This code takes longer to run, but it creates a table where each spaxel is a row. I do this to solve the problem of saving the table to a fits file. I also added a progress bar feature so I can visually see how far the code has run.